bindieselcurrent/requirements.txt
# Bin Diesel System Requirements

# Core dependencies
numpy>=1.21.0
python-dotenv>=1.0.0

# Wake word detection
pvporcupine>=3.0.0

# Object detection
ultralytics>=8.0.0

# Camera
# Install via: sudo apt install python3-picamera2

# GPIO control
RPi.GPIO>=0.7.0
# Alternative: gpiozero (if RPi.GPIO doesn't work)

# TOF Sensor (VL53L0X)
adafruit-circuitpython-vl53l0x>=2.4.0

# Voice recognition (manual mode)
openai>=1.0.0

# Audio input
pyaudio>=0.2.11

# OpenCV (for image processing)
# Install via: sudo apt install python3-opencv

